+++ scripts/importers/stats.js
@@ -11,12 +11,33 @@ const STAT_TYPE_MAP = {
     9: 'emp',
 }
 
-export async function updateStats(data, actor) {
-    const stats = data.stat.reduce((acc, stat) => {
-        const statName = STAT_TYPE_MAP[stat.stat_type_id]
-        acc[statName] = { value: stat.points };
+const STAT_NAME_MAP = {
+    "Intelligence": "int",
+    "Reflexes": "ref",
+    "Dexterity": "dex",
+    "Technique": "tech",
+    "Cool": "cool",
+    "Willpower": "will",
+    "Luck": "luck",
+    "Movement": "move",
+    "Body": "body",
+    "Empathy": "emp"
+}
+
+function getStatsV1(data) {
+    return data.stat.map((stat) => [STAT_TYPE_MAP[stat.stat_type_id], stat.points])
+}
+
+function getStatsV2(data) {
+    return Object.entries(data.stats).map(([statName, points]) =>
+        [STAT_NAME_MAP[statName], points])
+}
+
+export async function updateStats(data, actor, isV2) {
+    const stats = (isV2 ? getStatsV2(data) : getStatsV1(data)).reduce((acc, [statName, value]) => {
+        acc[statName] = { value };
         if (actor.system.stats[statName].max) {
-            acc[statName].max = stat.points;
+            acc[statName].max = value;
         }
         return acc;
     }, {});